filtros en historial de reportes
